[Feat] MockProtocol안에 enum제거 후 MockResponseType enum 추가
